horse-chess is done
This small exe can be divided into two parts. The first part is a welcome interface which need correct username and password(yyz 123)to successfully enter the main interface; The main face has five buttons(probably),can show the trace of the hourse on the chess board. Want more, download it and try!